move helper func out of model
